Fix some rare memory leaks and behaviour when a mapped window is reparented away from root
When a window that is mapped is reparented away from root, that is the same as it being completely unmapped+destroyed from our persective.  The window will no longer be in the same position on screen, and we will no longer want to listen to events on it.
Prior to this, the w->pixmap would not be freed when a mapped window was reparented away from root, and damage would not be updated for the screen correctly.