GitHub: Add an issue template
GitHub: Add a template for issues submitted, .github/issue_template.md.
I hope people could provide more needed information when reporting
issues after we provide a number of fields to fill and some extra
instructions in the template, to reduce the time wasted on the issues
because of insufficient information.

See: https://help.github.com/articles/creating-an-issue-template-for-your-repository/