Add a to_free member to reg_data_t
Xlib allocates separate memory for the rectangles and returns that to
the user. With xcb, instead one gets a pointer inside of the reply.
Thus, there, the rectangles are only valid until the reply is freed.

To prepare for this, this commit adds a to_free member to reg_data_t and
hopefully catches all places that need to be changed to this member.

Signed-off-by: Uli Schlachter <psychon@znc.in>