gl_common: set viewports and projection matrices when working with framebuffers and shaders When we are drawing a texture to a framebuffer, we must ensure that framebuffer's width and height are greater than or equal to texture's width and height, otherwise the texture will be cropped, so we are setting viewport to texture's width and height after binding a framebuffer.
When we are setting viewport to texture's width and height and using a
shader, we must update shader's projection matrix with corresponding
values (since projection matrices of all shaders are initialized with
framebuffer's default width and height), otherwise the result of using a
shader will look misplaced and/or scaled.

Fixes at least #212.