Add a .clang-tidy file
We currently don't use it for anything. We could add a CI pass in the
future.

We don't enable clang-analyzer checks. The checks enabled here should
mostly be syntax/style linting only, and should be cheap to run.
clang-analyzer checks are usually compute-intensive, and can easily be run
separately anyway, with scan-build and such.

The list of checks definitely isn't final.

Signed-off-by: Yuxuan Shui <yshuiv7@gmail.com>