Do not use single quotes in the example config
This replaces single quotes for double quotes in the configuration
comments that refer to strings.
This was misleading since someone could took that comment as a valid
config entry, which is not.  One notable example was `backend = 'glx'`
which is not accepted by libconfig.

This also adds `;` for each default config values for consistency.