added native optimization compiler flags
in my case building with `-O3 -march=native -mtune=native` gave me a 40% performance boost (not sure if it was due to flags or due to some new commits). Since most people will be building picom it for their personal use, it makes sense to leverage full potential of their CPUs.