Optimize animations so it doesn't do anything in case animation completed for a window.
The new animations kept on running even if there wasn't anything to do and so processor usage went up badly when animations turned on. This change checks if animation process completed for a window and in that case it simply jumps to the next window without doing all the calculations and other things.

For me this changed picom usage to the same as without animation so 50%+ cpu went down to ~9% just like without animations turned on from config.

Also because of this change animation_running=false is being used correctly as before this it was always true because in the for cycle it always set it to true even if it wasn't anything to do.