Reset velocity on animation end and animation_time
This should fix animations from starting at the wrong location as w->animation_velocity_... was never reset and because of the spring animation when a window was zoomed (for e.g. maximized and then restored) the next animation like shading became misplaced meaning it started at a wrong location because of the old velocity value.

Also we reset animation_time to 0 in case there was no animation running at all (just like it's for fading) and so we prevent for e.g. in case clamping turned off to start zooming in windows from thousands of times of your screen and freezing/braking everything.