Fixes black windows for Nvidia users
This should fix the black windows we were having in case of Firefox and other apps. It seems like only happened for Nvidia users as the pixmap was freed when binding new pixmaps but old pixmap wasn't freed and so the new binding didn't work and win_image ended up with a black pixmap.