Added possibility to define animation on window type basis
This will add possibility to define the animation on window type basis (which deprecated the animation-for-menu-window option) and so now you will be able to define animation like this:

wintypes:
{
  menu = { animation = "slide-down"; }
  popup_menu = { animation = "slide-down"; opacity = 0.97; }
  dropdown_menu = { animation = "slide-down"; opacity = 0.97; }
  notification = { animation = "slide-left"; }
  toolbar = { animation = "none"; }
};

The animation-for-transient-window has been kept as transient window isn't a type but a state so that's a separate thing.

NOTE: if "none" defined for animation for a specified window type that window won't be animated at all, the whole animation feature will be disabled for that kind of window so this way you can force specified windows to not animate at all for e.g. toolbars.