Better implementation of previous fix.
This fix for black windows is even better as it doesn't rely on animation progress and so it works even if animation disabled for those windows.
Tracked down the problem to this:

There's an ev_shape_notify() event fired quickly on new windows for e.g. in case of Firefox main menu and ev_shape_notify() sets the win_set_flags(w, WIN_FLAGS_SIZE_STALE); which brakes the new image captured and because this same event also sets w->reg_ignore_valid = false; too we check for this and this way we ignore the cloning of the image and so new image remains intact.