win: remove the SHADOW_* window flags
The window flags are used for delaying updates until we have grabbed the
X server. This is to make sure our internal states are in sync with the
server's state, to avoid race conditions.

Since shadow state is purely internal to picom, there is no need for
delaying its update, and thus these flags can be removed.

Signed-off-by: Yuxuan Shui <yshuiv7@gmail.com>