c2: remove match target format specifier
It's confusing to me what I should put as format, and the rule might or
might not work if I used the wrong format. In theory this can be used to
write condition that's only true when the specified property has a given
format, in reality I don't think anyone is doing that.

Signed-off-by: Yuxuan Shui <yshuiv7@gmail.com>