c2: warn if properties aren't used consistently
Make sure a property is not used both as a number and as a string. This
only apply to actual strings, atoms can be used as either a number or a
string.

Right now only a warning is printed, offending rules might or might not
work properly. But in the future these rules WILL be disabled and do
nothing.

Signed-off-by: Yuxuan Shui <yshuiv7@gmail.com>