backend: xrender: prefix backend-specific function names with xrender_
and sort the xrender_ops structure in the order backend operations are
declared in the backend_operations structure.

this makes the xrender backend's code consistent with the other
backends' code and reduces the risk of collisions because some backend
operation names are quite simple (e.g. init, deinit, fill, etc.).