dbus: make cdbus a bit more self-contained
cdbus_init/destroy no longer modifies struct session, instead they
returns the created struct cdbus_data and let the caller deal with it.

Also store the mainloop in cdbus_data, so the DBusWatch callbacks can
access it. We use to pass the whole `struct session` for the mainloop
and cdbus_data, but now we no longer set `cdbus_data` in `struct
session`, so we have to pass the `cdbus_data` as user data to the
DBusWatch callbacks, which means it has to contain the mainloop.

Signed-off-by: Yuxuan Shui <yshuiv7@gmail.com>