core: let's not do ancestry lookups
Always assume WM frames are direct children of the root window, and
client windows are direct children of their respective frames. Also WM
frames must have a one-to-one relationship with client windows.

We already assume this in most places, but it's inconsistent. Let's make
this assumption official. If something breaks we can then figure out a
principled way of dealing with that. Instead of having things "happen to
work" which is the situation now.

Signed-off-by: Yuxuan Shui <yshuiv7@gmail.com>